Separate activity from aws
