+++ service_activity_generated.go
@@ -1,0 +1,55 @@
+package main
+
+import (
+	"net/http"
+
+	"github.com/ca-risken/aws/proto/activity"
+)
+
+func (g *gatewayService) describeARNHandler(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	req := &activity.DescribeARNRequest{}
+	bind(req, r)
+	if err := req.Validate(); err != nil {
+		writeResponse(ctx, w, http.StatusBadRequest, map[string]interface{}{errorJSONKey: err.Error()})
+		return
+	}
+	resp, err := g.awsActivityClient.DescribeARN(ctx, req)
+	if err != nil {
+		writeResponse(ctx, w, http.StatusInternalServerError, map[string]interface{}{errorJSONKey: err.Error()})
+		return
+	}
+	writeResponse(ctx, w, http.StatusOK, map[string]interface{}{successJSONKey: resp})
+}
+
+func (g *gatewayService) listCloudTrailHandler(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	req := &activity.ListCloudTrailRequest{}
+	bind(req, r)
+	if err := req.Validate(); err != nil {
+		writeResponse(ctx, w, http.StatusBadRequest, map[string]interface{}{errorJSONKey: err.Error()})
+		return
+	}
+	resp, err := g.awsActivityClient.ListCloudTrail(ctx, req)
+	if err != nil {
+		writeResponse(ctx, w, http.StatusInternalServerError, map[string]interface{}{errorJSONKey: err.Error()})
+		return
+	}
+	writeResponse(ctx, w, http.StatusOK, map[string]interface{}{successJSONKey: resp})
+}
+
+func (g *gatewayService) listConfigHistoryHandler(w http.ResponseWriter, r *http.Request) {
+	ctx := r.Context()
+	req := &activity.ListConfigHistoryRequest{}
+	bind(req, r)
+	if err := req.Validate(); err != nil {
+		writeResponse(ctx, w, http.StatusBadRequest, map[string]interface{}{errorJSONKey: err.Error()})
+		return
+	}
+	resp, err := g.awsActivityClient.ListConfigHistory(ctx, req)
+	if err != nil {
+		writeResponse(ctx, w, http.StatusInternalServerError, map[string]interface{}{errorJSONKey: err.Error()})
+		return
+	}
+	writeResponse(ctx, w, http.StatusOK, map[string]interface{}{successJSONKey: resp})
+}
